refactor: Phase 3.1e — extract ResultsAndChartsMixin from ElectrodeAdvisorWidget
Move 13 results/chart/color methods (508 lines) into
ResultsAndChartsMixin in main_window_results_charts.py:
- _update_results_tables, _update_analysis_display
- _get_current_based_color, _get_path_current, _get_path_resistance
- _get_path_power, _get_path_temperature
- _calculate_color_scale_bounds, _value_to_color
- _draw_electrode_sphere, _update_temperature_profile
- _update_current_distribution, _update_power_distribution

Monolith now at 1,744 lines (-60.2% from original 4,387).

6 mixins extracted total:
- DrawingMixin: 510 lines (static geometry)
- InputPanelMixin: 402 lines (input controls)
- VisualControlsMixin: 485 lines (visual controls panel)
- VisualizationUpdateMixin: 934 lines (dynamic 3D updates)
- ResultsAndChartsMixin: 541 lines (results and chart updates)
